pref(relation): 删除遗留的grpc代码
