+++ internal/relation/api/grpc/v1/group_relation.proto
@@ -25,28 +25,6 @@ message UserIdsResponse {
   repeated string UserIds = 1;
 }
 
-
-message RemoveFromGroupRequest {
-  // @inject_tag: json:"group_id"
-  uint32 GroupId = 1;     // 群聊id
-  // @inject_tag: json:"user_id"
-  string UserId = 2;     // 移除的用户id
-}
-
-message RemoveFromGroupResponse {
-}
-
-message LeaveGroupRequest {
-  // @inject_tag: json:"group_id"
-  uint32 GroupId = 1;   // 群聊id
-  // @inject_tag: json:"user_id"
-  string UserId = 2;    // 拒绝的用户id
-}
-
-message LeaveGroupResponse {
-}
-
-
 message GetGroupRelationRequest {
   // @inject_tag: json:"group_id"
   uint32 GroupId = 1;     // 群聊id
@@ -111,25 +89,6 @@ message GetUserManageGroupIDResponse {
   repeated GroupIDRequest GroupIDs = 1;
 }
 
-message GetUserGroupIDsRequest {
-  // @inject_tag: json:"user_id"
-  string UserId = 1;  // 用户ID
-}
-
-message GetUserGroupIDsResponse {
-  // @inject_tag: json:"group_id"
-  repeated  uint32 GroupId = 1; // 群聊ID
-}
-
-message GetUserGroupListRequest {
-  // @inject_tag: json:"user_id"
-  string UserId = 1;
-}
-
-message GetUserGroupListResponse {
-  repeated UserGroupRequestList UserGroupRequestList = 1;
-}
-
 message UserGroupRequestList {
   // @inject_tag: json:"group_id"
   uint32 GroupId = 1;   // 群聊id
@@ -143,24 +102,6 @@ message UserGroupRequestList {
   int64 CreatedAt = 5;
   // @inject_tag: json:"inviter_id"
   string InviterId = 6;       // 邀请人id
-}
-
-message DeleteGroupRelationByGroupIdAndUserIDRequest {
-  // @inject_tag: json:"group_id"
-  uint32 GroupID = 1; // 群聊ID
-  // @inject_tag: json:"user_id"
-  string UserID = 2;  // 用户ID
-//  // @inject_tag: json:"status"
-//  GroupRelationStatus Status = 3;
-}
-
-message DeleteGroupRelationByGroupIdAndUserIDRevertRequest {
-  // @inject_tag: json:"group_id"
-  uint32 GroupID = 1; // 群聊ID
-  // @inject_tag: json:"user_id"
-  string UserID = 2;  // 用户ID
-//  // @inject_tag: json:"status"
-//  GroupRelationStatus Status = 3;
 }
 
 message CreateGroupAndInviteUsersRequest {
@@ -193,90 +134,17 @@ enum GroupSilentNotificationType {
   GroupSilent = 1;
 }
 
-message SetGroupSilentNotificationRequest{
-  // @inject_tag: json:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"group_id"
-  uint32 GroupId = 2;
-  // @inject_tag: json:"is_silent"
-  bool IsSilent = 3;
-}
-
 message CreateGroupAndInviteUsersResponse {
   // @inject_tag: json:"dialog_id"
   uint32 DialogId = 1;
-}
-
-message RemoveGroupRelationByGroupIdAndUserIDsRequest{
-  // @inject_tag: json:"group_id"
-  uint32 GroupId = 1;
-  // @inject_tag: json:"user_id"
-  repeated string UserIDs = 2;
-}
-
-message SetGroupOpenBurnAfterReadingRequest {
-  // @inject_tag: json:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"group_id"
-  uint32 GroupId = 2;
-  // @inject_tag: json:"open_burn_after_reading"
-  bool OpenBurnAfterReading = 3;
-}
-
-message SetGroupOpenBurnAfterReadingTimeOutRequest{
-  // @inject_tag: json:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"group_id"
-  uint32 GroupId = 2;
-  // @inject_tag: json:"open_burn_after_reading_time_out"
-  int64 OpenBurnAfterReadingTimeOut = 3;
-}
-
-message SetGroupUserRemarkRequest{
-  // @inject_tag: json:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"group_id"
-  uint32 GroupId = 2;
-  // @inject_tag: json:"remark"
-  string Remark = 3;
-}
-
-message AddGroupAdminRequest {
-  // @inject_tag: json:"group_id"
-  uint32 GroupID = 1;
-  string UserID = 2;
-  // @inject_tag: json:"user_ids"
-  repeated string UserIDs = 3;
-}
-
-message AddGroupAdminResponse {
-
 }
 
 service GroupRelationService {
   // 获取群聊成员ID列表
   rpc GetGroupUserIDs(GroupIDRequest) returns (UserIdsResponse);
 
-  // 获取用户的所有群聊ID列表
-  rpc GetUserGroupIDs(GetUserGroupIDsRequest) returns (GetUserGroupIDsResponse);
-
-  // 获取用户的所有群聊列表信息
-  rpc GetUserGroupList(GetUserGroupListRequest) returns (GetUserGroupListResponse);
-
-  // 获取群聊管理员ID列表
-  rpc GetGroupAdminIds(GroupIDRequest) returns (UserIdsResponse);
-
   // 获取用户管理的群聊ID列表
   rpc GetUserManageGroupID(GetUserManageGroupIDRequest) returns (GetUserManageGroupIDResponse);
-
-  // 从群聊中移除成员
-  rpc RemoveFromGroup (RemoveFromGroupRequest) returns (RemoveFromGroupResponse);
-
-  // 退出群聊
-  rpc LeaveGroup (LeaveGroupRequest) returns (LeaveGroupResponse);
-
-  // 退出群聊回滚操作
-  rpc LeaveGroupRevert(LeaveGroupRequest) returns (LeaveGroupResponse);
 
   // 获取用户与群聊关系信息
   rpc GetGroupRelation (GetGroupRelationRequest) returns (GetGroupRelationResponse);
@@ -290,33 +158,9 @@ service GroupRelationService {
   // DeleteGroupRelationByGroupId回滚操作
   rpc DeleteGroupRelationByGroupIdRevert(GroupIDRequest) returns (google.protobuf.Empty);
 
-  // 根据群聊ID和用户ID删除用户的群聊关系
-  rpc DeleteGroupRelationByGroupIdAndUserID(DeleteGroupRelationByGroupIdAndUserIDRequest) returns (google.protobuf.Empty);
-
-  // DeleteGroupRelationByGroupIdAndUserID回滚操作
-  rpc DeleteGroupRelationByGroupIdAndUserIDRevert(DeleteGroupRelationByGroupIdAndUserIDRequest) returns (google.protobuf.Empty);
-
   // 创建群聊并邀请多个用户进群（包括自己）
   rpc CreateGroupAndInviteUsers(CreateGroupAndInviteUsersRequest) returns (CreateGroupAndInviteUsersResponse);
 
   // CreateGroupAndInviteUsers 回滚操作
   rpc CreateGroupAndInviteUsersRevert(CreateGroupAndInviteUsersRequest) returns (google.protobuf.Empty);
-
-  // 设置群聊为静默通知状态
-  rpc SetGroupSilentNotification(SetGroupSilentNotificationRequest) returns (google.protobuf.Empty);
-
-  //移除多个用户
-  rpc RemoveGroupRelationByGroupIdAndUserIDs(RemoveGroupRelationByGroupIdAndUserIDsRequest) returns (google.protobuf.Empty);
-
-  // 设置阅后即焚开关
-  rpc SetGroupOpenBurnAfterReading(SetGroupOpenBurnAfterReadingRequest) returns (google.protobuf.Empty);
-
-  // 设置阅后即焚超时时间
-  rpc SetGroupOpenBurnAfterReadingTimeOut(SetGroupOpenBurnAfterReadingTimeOutRequest) returns (google.protobuf.Empty);
-
-  //设置群聊用户备注
-  rpc SetGroupUserRemark(SetGroupUserRemarkRequest) returns (google.protobuf.Empty);
-
-  // 添加群聊管理员
-  rpc AddGroupAdmin(AddGroupAdminRequest) returns (AddGroupAdminResponse);
 }
