+++ internal/relation/api/grpc/v1/user_relation.proto
@@ -1,50 +1,16 @@
 syntax = "proto3";
 package relation_v1;
 option go_package = "github.com/cossim/coss-server/internal/relation/api/grpc/v1";
-import "google/protobuf/empty.proto";
 
 service UserRelationService {
   // 添加好友
   rpc AddFriend(AddFriendRequest) returns (AddFriendResponse);
-  // 管理好友
-  rpc ManageFriend(ManageFriendRequest) returns (ManageFriendResponse);
-  // 管理好友回滚
-  rpc ManageFriendRevert(ManageFriendRequest) returns (ManageFriendResponse);
-  // 删除好友
-  rpc DeleteFriend(DeleteFriendRequest) returns (DeleteFriendResponse);
-  // 删除好友回滚
-  rpc DeleteFriendRevert(DeleteFriendRequest) returns (DeleteFriendResponse);
-  // 添加黑名单
-  rpc AddBlacklist(AddBlacklistRequest) returns (AddBlacklistResponse);
-  // 移出黑名单
-  rpc DeleteBlacklist(DeleteBlacklistRequest) returns (DeleteBlacklistResponse);
   // 获取好友列表
   rpc GetFriendList(GetFriendListRequest) returns (GetFriendListResponse);
-  // 获取黑名单列表
-  rpc GetBlacklist(GetBlacklistRequest) returns (GetBlacklistResponse);
   // 获取用户关系
   rpc GetUserRelation(GetUserRelationRequest) returns (GetUserRelationResponse);
   // 批量获取用户关系
-  rpc GetUserRelationByUserIds(GetUserRelationByUserIdsRequest) returns (GetUserRelationByUserIdsResponse);
-  // 设置好友静默通知
-  rpc SetFriendSilentNotification(SetFriendSilentNotificationRequest) returns (google.protobuf.Empty);
-  // 设置阅后即焚开关
-  rpc SetUserOpenBurnAfterReading(SetUserOpenBurnAfterReadingRequest) returns (google.protobuf.Empty);
-  // 设置好友备注
-  rpc SetFriendRemark(SetFriendRemarkRequest) returns (google.protobuf.Empty);
-  // 单向删除好友后再添加回好友
-  rpc AddFriendAfterDelete(AddFriendAfterDeleteRequest) returns (AddFriendAfterDeleteResponse);
-}
-
-message AddFriendAfterDeleteRequest {
-  // @inject_tag: json:"user_id"
-  string UserId = 1;    // 用户id
-  // @inject_tag: json:"friend_id"
-  string FriendId = 2;  // 好友id
-}
-
-message AddFriendAfterDeleteResponse {
-
+  rpc GetRelationsWithUsers(GetUserRelationByUserIdsRequest) returns (GetUserRelationByUserIdsResponse);
 }
 
 message AddFriendRequest {
@@ -56,54 +22,6 @@ message AddFriendRequest {
 }
 
 message AddFriendResponse {
-  // Add any necessary fields here for the response
-}
-
-message ManageFriendRequest {
-  // @inject_tag: json:"user_id" form:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"friend_id" form:"friend_id"
-  string FriendId = 2;
-  // @inject_tag: json:"dialog_id" form:"dialog_id"
-  uint32 DialogId = 3;
-  // @inject_tag: json:"status" form:"status"
-  RelationStatus Status = 4;
-}
-
-message ManageFriendResponse {
-  // Add any necessary fields here for the response
-}
-
-message DeleteFriendRequest {
-  // @inject_tag: json:"user_id" form:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"friend_id" form:"friend_id"
-  string FriendId = 2;
-}
-
-message DeleteFriendResponse {
-  // Add any necessary fields here for the response
-}
-
-message AddBlacklistRequest {
-  // @inject_tag: json:"user_id" form:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"friend_id" form:"friend_id"
-  string FriendId = 2;
-}
-
-message AddBlacklistResponse {
-  // Add any necessary fields here for the response
-}
-
-message DeleteBlacklistRequest {
-  // @inject_tag: json:"user_id" form:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"friend_id" form:"friend_id"
-  string FriendId = 2;
-}
-
-message DeleteBlacklistResponse {
   // Add any necessary fields here for the response
 }
 
@@ -133,21 +51,6 @@ message GetFriendListRequest {
 message GetFriendListResponse {
   // @inject_tag: json:"friend_list" form:"friend_list"
   repeated Friend FriendList = 1;
-}
-
-message Blacklist {
-  // @inject_tag: json:"user_id" form:"user_id"
-  string UserId = 1;
-}
-
-message GetBlacklistRequest {
-  // @inject_tag: json:"user_id" form:"user_id"
-  string UserId = 1;
-}
-
-message GetBlacklistResponse {
-  // @inject_tag: json:"blacklist" form:"blacklist"
-  repeated Blacklist Blacklist = 1;
 }
 
 message GetUserRelationRequest {
@@ -192,41 +95,4 @@ message GetUserRelationByUserIdsRequest {
 
 message GetUserRelationByUserIdsResponse {
   repeated GetUserRelationResponse Users = 1;
-}
-
-enum UserSilentNotificationType {
-  // @inject_tag: json:"not_silent"
-  UserNotSilent = 0;
-  // @inject_tag: json:"silent"
-  UserSilent = 1;
-}
-
-message SetFriendSilentNotificationRequest {
-  // @inject_tag: json:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"friend_id"
-  string FriendId = 2;
-  // @inject_tag: json:"is_silent"
-  bool IsSilent = 3;
-}
-
-message SetUserOpenBurnAfterReadingRequest {
-  // @inject_tag: json:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"friend_id"
-  string FriendId = 2;
-  // TimeOut 阅后即焚时间
-  // @inject_tag: json:"timeout"
-  int64 TimeOut = 3;
-  // @inject_tag: json:"open_burn_after_reading"
-  bool OpenBurnAfterReading = 4;
-}
-
-message SetFriendRemarkRequest {
-  // @inject_tag: json:"user_id"
-  string UserId = 1;
-  // @inject_tag: json:"friend_id"
-  string FriendId = 2;
-  // @inject_tag: json:"remark"
-  string remark = 3;
 }
