--- internal/relation/service/dtm_workflow.go
@@ -1,1 +1,0 @@
-package service
